FAIMS-245: fixing bug where finally block is not called in services and client which cleans up temp files

+++ faims-android-app/src/au/org/intersect/faims/android/util/FileUtil.java
@@ -286,6 +286,9 @@ public class FileUtil {
 	}
 
 	public static void deleteDirectory(File dir) {
+		if (!dir.isDirectory()) {
+			return;
+		}
 		for (File f : dir.listFiles()) {
 			if (f.isDirectory()) {
 				deleteDirectory(f);
